Refactor : Script 파일 구조 변경 및 모듈화
