+++ src/pages/account/signin/index.html
@@ -58,5 +58,5 @@
       </form>
     </div>
   </body>
-  <script type="module" src="../account.js"></script>
+  <script type="module" src="signin.js"></script>
 </html>
