+++ src/pages/account/signup/index.html
@@ -65,5 +65,5 @@
       </form>
     </div>
   </body>
-  <script type="module" src="../account.js"></script>
+  <script type="module" src="signup.js"></script>
 </html>
